colocado erro handling e verificaçao de cliente vip na venda nova classe verificarpontos mudança na interface promoçao remocao descontoclientevip descontoclientevip foi de public para private

+++ src/br/edu/cafeteria/excecao/PontosInsuficientesException.java
@@ -1,16 +1,15 @@
 package br.edu.cafeteria.excecao;
 
+import br.edu.cafeteria.modelo.Cliente;
 public class PontosInsuficientesException extends Exception {
 	
 	private final float pontosNecessarios;
 	private final float pontosDisponiveis;
-	
-	public PontosInsuficientesException(float pontosNecessarios, float pontosDisponiveis) {
+	public PontosInsuficientesException(int pontosNecessarios, int pontosDisponiveis) {
 		super("Pontos de XP insuficientes: necessario " + pontosNecessarios + ", disponivel " + pontosDisponiveis);
 		this.pontosNecessarios = pontosNecessarios;
 		this.pontosDisponiveis = pontosDisponiveis;
 	}
-
 	public float getPontosNecessarios() {
 		return pontosNecessarios;
 	}
